feat: localize common utilities (errors, routes)
