+++ packages/abstract-provider/src/errors.ts
