+++ packages/abstract-provider/src/index.ts
@@ -1,1 +1,3 @@
 export * from './abstract-provider'
+export * from './errors'
+export * from './routes'
